brew_cask_update works reliably

+++ fish/functions/brew_update_casks.fish
@@ -4,7 +4,9 @@ function brew_update_casks --description 'update outdated casks'
   set -x _outdated_cask_names (echo $_outdated_casks | gsed "s/ /\n/g")
 
   for val in $_outdated_cask_names
-    rm -rf /opt/homebrew-cask/Caskroom/$val
-    brew cask install $val
+    if [ $val ]
+      rm -rf /opt/homebrew-cask/Caskroom/$val
+      brew cask install $val
+    end
   end
 end
